Fixed implementation of complex number square root
Revised implementation and added a new test

+++ src/pow.rs
@@ -1,12 +1,12 @@
 pub trait Pow {
-    fn powf(&self, n: Self) -> Self;
+    fn pow(&self, n: Self) -> Self;
 }
 
 macro_rules! impl_pow {
     ( $($ty:ty),* ) => {
         $(
             impl Pow for $ty {
-                fn powf(&self, n: Self) -> Self {
+                fn pow(&self, n: Self) -> Self {
                     self.powf(n)
                 }
             }
